Calls to bk_printf with logging strings need %s
bk_printf expects a formatting string, so if you pass it a log entry with a "%" in it, it will not show the "%" symbol and instead try to match a formatting pattern. This causes a difference in logging to serial vs logging to TCP / HTTP. Eg, this code results in different results on serial vs other logging:

```
ADDLOGF_INFO("At %d%%", 12);
```

Expected: Log shows "At 12%" - actual serial log is "At 12".