Make LFS size adjustable.  Breaking change as default is 0x8000 (16Kbytes). New commands: lfssize - log size lfssize 0x10000 - configure size lfsunmount lfsmount lfsformat - reformat at current size lfsformat 0x18000 - configre size and format lfs.
lfs size is stored in devcie config.
The main benefit of this is if a small LFS is used, it does not get overwritten when doing OTA.
Also, if it DOES, you can still backup and restore, which will be faster with a smaller area.
You still have the option for up to 0x80000 - 512 kbytes.