allow storage of WPA2 PSK as hexadecimal string (#594)
IEEE 802.11i limits passwords to 63 characters to allow them to be
distinguished from PSKs in 64 hexadecimal digit form. To support 64
character strings without changing the layout of the configuration
structure in flash, this commit drops the requirement that there has
to be a terminating nul byte inside the array. Without changing the
API this comes at the cost of an additional 65 byte buffer in RAM.