tuyaMCU: avoid state machine interfering with reset AP mode (#1597)
Don't continue with the state machine if the device is in Open WiFi mode
based on a user button press.

This avoids the device shutting down itself when in recovery mode. The
MCU will reset the device after 90 seconds automatically.

Communicate to the MCU the device is in configuration mode. This will
change the status LED blink pattern to reflect the state.

Signed-off-by: David Bauer <mail@david-bauer.net>