Add file with all role related data and an awk script to generate "src/rolesNchannels.h". This will be imported by all files using pin-roles and its attributes (i.e. number of channels) while "old" information is commented out.
The input file also includes information on use of channels ("T" for temperature, "H" for humidity ...).

If you set
it will include this information to pin config page.

Size of resulting image is slightly smaller due to som optimizations in JS code.

If you include the additional information, it will be approx 350bytes larger